refactor: extract edit cv work experience section to own component
