+++ cv/src/components/CvEditForm.jsx
@@ -3,6 +3,7 @@ import "../styles/CvEditForm.css"
 import Experience from "../model/Experience"
 import CvEditAchievement from "./CvEditAchievement"
 import CvEditEducation from "./CvEditEducation"
+import CvEditWorkExperience from "./CvEditWorkExperience"
 
 const nameValidation = {
   required: "Name is required",
@@ -120,33 +121,18 @@ export default function CvEditForm({ cvDetails, handleCvSubmit }) {
             {errors.contactSummary.message}
           </span>
         )}
-        <div
-          className="edit-cv-work-experience-container"
-          data-cy="edit-cv-work-experience-container"
-        >
-          <p className="section-title">Work Experience</p>
-          {errors.workExperiences && errors.workExperiences.empty ? (
-            <p className="error" data-cy="error-work-experience-empty">
-              {errors.workExperiences.empty.message}
-            </p>
-          ) : null}
-          {getWorkExperienceSection(workExperiencesFields, {
-            errors,
+        <CvEditWorkExperience
+          fields={workExperiencesFields}
+          formFunctions={{
             register,
+            errors,
             removeWorkExperience,
-          })}
-          <button
-            type="button"
-            onClick={() => {
+            handleAddWorkExperience: () => {
               appendWorkExperience(new Experience("", "", ""))
               clearErrors("workExperiences.empty")
-            }}
-            className="add-work-experience-btn"
-            data-cy="edit-cv-add-work-experience-btn"
-          >
-            + Add Work Experience
-          </button>
-        </div>
+            },
+          }}
+        />
         <CvEditEducation
           fields={educationFields}
           formFunctions={{
@@ -160,7 +146,7 @@ export default function CvEditForm({ cvDetails, handleCvSubmit }) {
           }}
         />
         <CvEditAchievement
-          achievementFields={achievementFields}
+          fields={achievementFields}
           formFunctions={{
             register,
             errors,
@@ -180,92 +166,4 @@ export default function CvEditForm({ cvDetails, handleCvSubmit }) {
       </form>
     </div>
   )
-}
-
-const workExperienceTitleValidation = {
-  required: "Title is required",
-  maxLength: {
-    value: 100,
-    message: "Work experience title cannot be more than 100 characters",
-  },
-}
-const workExperienceDescriptionValidation = {
-  required: "Description is required",
-  maxLength: {
-    value: 300,
-    message: "Work experience description cannot be more than 300 characters",
-  },
-}
-
-function getWorkExperienceSection(
-  workExperiencesFields,
-  { errors, register, removeWorkExperience }
-) {
-  const lastElementIndex = workExperiencesFields.length - 1
-  return workExperiencesFields.map((field, index) => {
-    return (
-      <section key={field.id} className="work-experience-card">
-        <label htmlFor={`workExperiences.${index}.title`}>Title</label>
-        <input
-          key={`${field.id}-${index}-title`}
-          type="text"
-          data-cy={`edit-cv-work-experience-${index}-title`}
-          {...register(
-            `workExperiences.${index}.title`,
-            workExperienceTitleValidation
-          )}
-        />
-        {errors.workExperiences &&
-          errors.workExperiences[index] &&
-          errors.workExperiences[index].title && (
-            <span
-              className="error"
-              data-cy={`edit-work-experience-title-${index}-error`}
-            >
-              {errors.workExperiences[index].title.message}
-            </span>
-          )}
-
-        <label htmlFor={`workExperiences.${index}.description`}>
-          Description
-        </label>
-        <textarea
-          key={`${field.id}-${index}-description`}
-          data-cy={`edit-work-experience-${index}-description`}
-          rows={5}
-          {...register(
-            `workExperiences.${index}.description`,
-            workExperienceDescriptionValidation
-          )}
-        />
-        {errors.workExperiences &&
-          errors.workExperiences[index] &&
-          errors.workExperiences[index].description && (
-            <span
-              className="error"
-              data-cy={`edit-work-experience-description-${index}-error`}
-            >
-              {errors.workExperiences[index].description.message}
-            </span>
-          )}
-
-        <label htmlFor={`workExperiences.${index}.dateRange`}>Date Range</label>
-        <input
-          key={`${field.id}-${index}-dateRange`}
-          type="text"
-          data-cy={`edit-work-experience-${index}-dateRange`}
-          {...register(`workExperiences.${index}.dateRange`)}
-        />
-        <button
-          type="button"
-          onClick={() => removeWorkExperience(index)}
-          className="edit-cv-work-experience-delete-btn"
-          data-cy={`edit-cv-work-experience-${index}-delete-btn`}
-        >
-          Delete Work Experience
-        </button>
-        {index === lastElementIndex ? null : <hr className="divider" />}
-      </section>
-    )
-  })
 }
